refactor: rename service implementations for consistency in package structure
